# docs/swagger.yaml
basePath: /
definitions:
...
  model.Department:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.Permission:
    properties:
      action:
        description: view, create, update, delete
        type: string
      feature_flag:
        description: tên hoặc mã feature flag, rỗng nếu không dùng
        type: string
      id:
        type: integer
      resource:
        $ref: '#/definitions/model.Resource'
      resource_id:
        type: integer
      roles:
        items:
          $ref: '#/definitions/model.Role'
        type: array
    type: object
  model.Position:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.Resource:
    properties:
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/model.Permission'
        type: array
    type: object
  model.Role:
    properties:
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/model.Permission'
        type: array
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.Stack:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.User:
    properties:
      department:
        $ref: '#/definitions/model.Department'
      department_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      position:
        $ref: '#/definitions/model.Position'
      position_id:
        type: integer
      role:
        $ref: '#/definitions/model.Role'
      role_id:
        type: integer
      stack:
        $ref: '#/definitions/model.Stack'
      stack_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for a Gin application.
  title: My Gin App API
  version: "1.0"
paths:
...
  /feature-flags:
    get:
      description: get all feature flags (distinct from permissions)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - BearerAuth: []
      summary: List feature flags
      tags:
      - feature-flags
    post:
      consumes:
      - application/json
      description: create a new feature flag (as a permission with feature_flag field)
      parameters:
      - description: Feature flag info
        in: body
        name: flag
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Permission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create feature flag
      tags:
      - feature-flags
...
  /permissions:
    get:
      description: get all permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Permission'
            type: array
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - permissions
    post:
      consumes:
      - application/json
      description: create a new permission
      parameters:
      - description: Permission to create
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/model.Permission'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Permission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create permission
      tags:
      - permissions
...
  /resources:
    get:
      description: get all resources
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Resource'
            type: array
      security:
      - BearerAuth: []
      summary: List resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: create a new resource
      parameters:
      - description: Resource to create
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/model.Resource'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Resource'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create resource
      tags:
      - resources
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"